pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
authlib = "0.13"
httpx = "0.7.8"
uvicorn = "^0.15.0"
fastapi = "^0.70.0"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.26.0"
tenacity = "^6.1.0"
pydantic = {extras = ["email"], version = "^1.3"}
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.1.0"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
pytest = "^5.4.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
aiofiles = "^0.7.0"
PyJWT = "^2.2.0"
starlette = "^0.16.0"
fastapi-utils = "^0.2.1"
itsdangerous = "^2.0.1"
pymongo = "^3.11.1"
motor = "^2.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

